+++ aksel.nav.no/website/pages/admin/[[...index]].js
@@ -15,7 +15,6 @@ const StudioPage = () => {
         config={workspaceConfig}
         scheme={scheme}
         onSchemeChange={setScheme}
-        sche
         unstable__noFavicons
       />
     </div>
@@ -23,3 +22,10 @@ const StudioPage = () => {
 };
 
 export default StudioPage;
+
+/*
+TODO:
+Sort-order basert på siste godkjent/utdatert? publisedAt?
+Action for publisedAt ved første publisering
+Legge til "godkjent"-action på grunnleggende, komponenter og god-praksis
+*/