+++ aksel.nav.no/website/sanity/custom-plugins/structure/god-praksis.ts
@@ -1,3 +1,5 @@
+import { FileError } from "@navikt/ds-icons";
+import { MasterDetailIcon } from "@sanity/icons";
 import { StructureBuilder } from "sanity/desk";
 /* documentStore is in Alpha, so avoid using for now */
 export const GodPraksisPanes = async (getClient, S: StructureBuilder) => {
@@ -25,6 +27,7 @@ export const GodPraksisPanes = async (getClient, S: StructureBuilder) => {
   return [
     S.listItem()
       .title(`Temasider (${tema.length})`)
+      .icon(MasterDetailIcon)
       .child(S.documentTypeList("aksel_tema")),
     S.divider(),
     ...tema.map(({ title, _id }) =>
@@ -45,6 +48,7 @@ export const GodPraksisPanes = async (getClient, S: StructureBuilder) => {
     S.divider(),
     S.listItem()
       .title(`Artikler uten tema (${ids.filter((x) => !x?.tema).length ?? 0})`)
+      .icon(FileError)
       .child(
         S.documentList()
           .title(`Artikler uten tema`)