+++ aksel.nav.no/website/sanity/schema/documents/prinsipper/landingsside.tsx
@@ -1,3 +1,4 @@
+import { HomeIcon } from "@sanity/icons";
 import { defineField, defineType } from "sanity";
 
 export const PrinsipperLandingSide = defineType({
@@ -14,6 +15,7 @@ export const PrinsipperLandingSide = defineType({
   preview: {
     prepare: () => ({
       title: "Landingsside prinsipper",
+      media: HomeIcon,
     }),
   },
 });