+++ @navikt/internal/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @navikt/ds-react-internal
 
+## 2.0.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @navikt/ds-react@2.0.10
+  - @navikt/ds-icons@2.0.10
+
 ## 2.0.9
 
 ### Patch Changes